Assert: Norm
