+++ FL_TESTS
@@ -3,6 +3,7 @@ test/AssertConstructor.h
 test/AssertDefuzzifier.h
 test/AssertFactory.h
 test/AssertHedge.h
+test/AssertNorm.h
 test/Headers.h
 test/Listener.h
 test/Mock.h
